test: add tests for integrate_sin and integrate_cos
This extracts the duplicated testing code for `integrate_sin` and
`integrate_cos` out of the monolithic `test_Calculus.py` test suite
into a dedicated `test_TrigIntegration.py` suite.

The new test suites provides complete functional line and logic
coverage over the `Math.Calculus.Integration.TrigIntegration` module.
